chore: moved util functions to their own file
